+++ src/models/paymentOrders/request models/paymentOrderCancelRequest.ts
@@ -1,4 +1,4 @@
-import { PaymentOrderCancelTransaction } from "../paymentOrderCancelTransaction";
+import { PaymentOrderCancelTransaction } from "./paymentOrderCancelTransaction";
 
 export interface PaymentOrderCancelRequest {
     transaction: PaymentOrderCancelTransaction;
