+++ src/models/paymentOrders/request models/paymentOrderCaptureRequest.ts
@@ -1,4 +1,4 @@
-import { PaymentOrderCaptureTransaction } from "../paymentOrderCaptureTransaction";
+import { PaymentOrderCaptureTransaction } from "./paymentOrderCaptureTransaction";
 
 export interface PaymentOrderCaptureRequest {
     transaction: PaymentOrderCaptureTransaction;
